refactor: extract ChunkGenerator from ChunkManager
Moves all procedural generation out of ChunkManager into a new
ChunkGenerator class: chunk contents (_generateChunk), star/planet type
selection, position scoring, and the eleven per-type generators
including the cross-chunk wormhole pair store.

Generation code is moved verbatim - seed derivations and RNG call order
are untouched, so existing universes and shared seeds are unaffected.
The generator reads the world through a narrow WorldContext interface
(chunk cache, discovery state, debug spawns, region queries) via
accessor mirrors, which is what allowed a zero-edit move.

- ChunkManager: 1,818 -> 451 lines; now purely chunk lifecycle,
  region queries, and API delegation
- Removes dead _createEmptyChunk (defined, never called)
- Adds determinism lock tests (independent managers generate identical
  chunks) verified against the old implementation first
